third_party/WebKit/LayoutTests/fast/dom/Range/31684-expected.txt

This test checks an orphan text node cannot be surrounded by the range. (bug31684)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.surroundContents(element) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS successfullyParsed is true

TEST COMPLETE
